pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-eval"
version = "0.1.0"
description = "Claude + Wikipedia question-answering with an eval harness"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
wiki-eval = "wiki_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wiki_eval = ["prompts/*.md"]
